requirements.txt
python-dotenv
pinecone-client
llama-hub
PyMuPDF
openai
transformers